完成对话

+++ include/ubuf/ubuf_const.h
@@ -2,7 +2,7 @@
 
 #define UBUF_CLONE_STRING(pucSrcString, pstDstUbuf, pucDstString) \
 {\
-    ULONG ulLen = strlen(pucSrcString);\
+    ULONG ulLen = (ULONG)strlen(pucSrcString);\
     pucDstString = UBUF_AddComponent(pstDstUbuf, ulLen+1);\
     strcpy(pucDstString, pucSrcString);\
     pucDstString[ulLen] = '\0';\
